resource-timing/object-reload.sub.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing object reload</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    let t = async_test("Test that object reloads are not observable by the parent");
    function verify_iframe_reload_not_observable() {
        let entries = performance.getEntriesByType("resource");
        let found_first_iframe = false;
        for (entry of entries) {
            if (entry.name == "{{location[scheme]}}://{{host]}}:{{ports[http][0]}}/resource-timing/resources/iframe-that-reloads.html") {
                found_first_iframe = true;
            }
            assert_not_equals(entry.name, "{{location[scheme]}}://{{host]}}:{{ports[http][0]}}/resource-timing/resources/reloaded_iframe.html", "Reloaded object should not be observable");
        }
        assert_true(found_first_iframe, "Initial object should be observable");
        t.done();
    }
    window.addEventListener("message", t.step_func(e=>{
        if (e.data == "reloaded") {
            verify_iframe_reload_not_observable();
        }
    }));
</script>
<object type="text/html" data="{{location[scheme]}}://{{host]}}:{{ports[http][0]}}/resource-timing/resources/iframe-that-reloads.html">
